Read in and sort PGS data

Given the path containing PGS data, this function will read in PGS data for the PGS's specified and the significance levels provided. In order to correctly only keep rows of data from verified sources, the genetic_match_file is necessary to provide for a check.

get_pgs(pgs = c("AD", "AD_Jansen"), s_levels = c("S1", "S7", "S11"),
  pgs_path = "~/LCBC/Projects/Cross_projects/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
  genetic_match_file = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
  include_cnt = FALSE, include_genetics_debug = FALSE)

Arguments

pgs
character vector of the PGS wanted

s_levels
character vector of the significance levels wanted

pgs_path
path to the directory where the PGS's are stored

genetic_match_file
path to the file containing the MOAS-genetics matching and debugging information

include_cnt
logical, whether to keep SNP count information

include_genetics_debug
logical, whether to keep all columns in the genetic_match_file in the final output

Value

a tibble / data.frame

Examples

# NOT RUN {
# Here we assume you know the path to the LCBC
# lagringshotell, you can substitute "~" in the
# paths with the path to the lagringshotell

get_pgs( pgs = c("AD", "EduYears_2016", "Depression_Nagel2018"),
         s_levels = c("S1", "S7", "S11"),
         pgs_path = "~/LCBC/Projects/Cross_projects/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
         genetic_match_file = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
)

# You can also toggle adding the CNT columns from the PGS, by changing
# include_cnt to TRUE
get_pgs( pgs = c("AD", "EduYears_2016", "Depression_Nagel2018",
         s_levels = c("S1", "S7", "S11"),
         pgs_path = "~/LCBC/Projects/Cross_projects/Genetics/PGS/PGS_20190618/PGS_wAPOE/",
         genetic_match_file = "~/LCBC/Projects/Cross_projects/MOAS/data-raw/DNA/gID_MOAS_match.tsv",
         include_cnt = FALSE
)

# }